Update the name of the type alias of the GraphQLObject
Update the name of the type alias of the GraphQLObject from `Query`
to `_Query`. This was done since it was creating a name collision in
swift if the introspection schema had a type named `Query`.

Note: The problem will still happen if for some reason there is a type
named `_Query` in the introspection schema